add float32 functionality

+++ npbench/benchmarks/polybench/covariance2/covariance2.py
@@ -3,7 +3,7 @@
 import numpy as np
 
 
-def initialize(M, N, datatype=np.float64):
+def initialize(M, N, datatype=np.float32):
     float_n = datatype(N)
     data = np.fromfunction(lambda i, j: (i * j) / M, (N, M), dtype=datatype)
