+++ npbench/benchmarks/polybench/k3mm/k3mm.py
@@ -3,7 +3,7 @@
 import numpy as np
 
 
-def initialize(NI, NJ, NK, NL, NM, datatype=np.float64):
+def initialize(NI, NJ, NK, NL, NM, datatype=np.float32):
     A = np.fromfunction(lambda i, j: ((i * j + 1) % NI) / (5 * NI), (NI, NK),
                         dtype=datatype)
     B = np.fromfunction(lambda i, j: ((i * (j + 1) + 2) % NJ) / (5 * NJ),
